Fix `InvocationTargetException` in ScalaDoc gen
This sbt command was failing with an `InvocationTargetException`:

```
+aws-s3-sdk-v2 / Compile / doc
```

...the 'doc' command is invoked during release, so we need it to work!

This is a known issue: https://github.com/lampepfl/dotty/issues/15288

See also https://github.com/scanamo/scanamo/pull/1643, which looks like
it fixed the issue by downgrading to Scala 3.2.0. Weirdly, this fix did
not seem to work for me when I tried it here.