Add stub script for compatibility-labelling PRs
This is a stub script that doesn't actually output true compatibility
information, it's just a demonstration showing what kind of behaviour
would be desirable.

It's aiming to show the part 1 of the workflow described in
https://github.com/scalacenter/sbt-version-policy/pull/179:

> **Keep devs aware of the compatibility of the PR they're working on** : When a PR is opened on a library repository, an automated workflow comments-on or labels the PR, making them aware that _"This PR breaks binary/source compatibility"_ - the information here would come from the `sbt-version-policy` plugin, but wouldn't require setting a `versionPolicyIntention` - it just establishes _what_ the level of compatibility the PR entails, so that the developer is aware. The assumption here is that whatever the developer is doing, they need to do it, and should just know the compatibility cost the PR is going to have (and possibly, though not necessarily, once informed they may be able to modify the PR to improve the compatibility).

The script is written to add-and-remove labels as appropriate, if
the required compatibility label (eg 'Breaks Binary Compatibility') wasn't already
present, we assume this is new information and leave a detailed comment.
In general, to reduce noise, we wouldn't necessarily want to add a new
comment on every push, better to just tell the user how they can get this
information by running locally...