Suppress Caffeine exceptions: model `Missing` values
We want a way for our
`Loading` object to indicate the value for a key is `Missing` (eg we are
trying to load an S3 key that does not exist) without having to throw an
exception - as Caffeine will noisily log almost every exception it sees,
even if the client does not mind missing values, as is the case with
the Facia Scala Client.

We've introduced the `MissingOrETagged` sealed trait to allow us to
represent that a fetch result may be either missing or found, returned
with ETagged data.

This is an alternative approach to the hacky solution in
https://github.com/guardian/etag-caching/pull/32 .

Co-authored-by: Jamie B <53781962+JamieB-gu@users.noreply.github.com>